// types/event.d.ts
export namespace EventsApi {
    interface TicketingAware {
        offersUrl?: Array<string>
    }

    interface DocumentedEntity {
        mainDocument?: RoadizDocument | null
        mainDocuments?: Array<RoadizDocument> | null
        medias?: Array<RoadizDocument>
        logos?: Array<RoadizDocument>
        icons?: Array<RoadizDocument>
        fallbackDocument?: RoadizDocument | null
    }

    interface TaggedEntity {
        tags?: Array<RoadizTag> // collection without filter
        mainTag?: RoadizTag | null // first main tag
        mainTags?: Array<RoadizTag> | null
        otherTags?: Array<RoadizTag> | null
    }

    interface Thing {
        name?: string | null
        slug?: string | null
    }

    interface Routable {
        url?: string | null
    }

    interface PublishableEntity {
        publishedAt?: string | null
        published?: boolean
    }

    interface DateSortable {
        sortingFirstDateTime?: string | null
        sortingDateTime?: string | null
        sortingLastDateTime?: string | null
        availability?: number
        cancelled?: boolean
        postponed?: boolean
        programmingInProgress?: boolean
        arrayDates?: Array<string> | Array<Array<string>>
        longRunningEvent?: boolean
        placesNames?: Record<string, string>
    }

    interface Event
        extends JsonLdObject,
            Thing,
            Routable,
            DocumentedEntity,
            TicketingAware,
            TaggedEntity,
            DateSortable,
            PublishableEntity {
        additionalRoles?: string | null
        attributes?: Array<KeyValueItem>
        blogPosts?: Array<RoadizNodesSources>
        bookingDescription?: string | null
        bookingLabel?: string | null
        castingPending?: boolean
        color?: string | null
        country?: string | null
        customDatesLabel?: string | null
        dates: Array<Omit<EventDate, 'event' | 'additionalTicketingData' | 'offers'>>
        datesCount?: number | null
        disclaimer?: string | null
        duration?: string | null
        excerpt?: string | null
        gatheredEvents?: Array<Event>
        gatheredEventsCount?: number | null
        inLanguage?: string | null
        includedIn?: Array<Event>
        minPrice?: number | null
        maxPrice?: number | null
        offSubscription?: boolean
        placeMapImage?: RoadizDocument | null
        otherDocuments?: Array<RoadizDocument>
        overtitleLanguage?: string | null
        peripheralEvents?: Array<Event>
        position?: number | null
        priceRange?: string | null
        pricingDescription?: string | null
        reassurance?: Review | null
        relatedEvents?: Array<Event>
        reversePeripheralEvents?: Array<Event>
        reverseRelatedEvents?: Array<Event>
        review?: Review | null
        performers?: Array<Omit<Role, 'event'>>
        sponsors?: Array<Omit<Role, 'event'>>
        scheduleDescription?: string | null
        schedule?: Array<ScheduleItem>
        season?: Omit<Season, 'seo'> | null
        seo?: Seo | null
        sponsorsTitle?: string | null
        subEvents?: Array<Event>
        subEventsCount?: number | null
        subtitle?: string | null
        superEvent?: Event | null
        type?:
            | null
            | 'live_performance'
            | 'festival'
            | 'educational'
            | 'exhibition'
            | 'screening'
            | 'conference'
            | 'guided_tour'
            | 'piece_of_art'
            | 'other'
        typicalAgeRange?: string | null
        locale?: string | null
    }
// ...
    interface EventDate extends JsonLdObject, DurationAware, Availability {
        event?: Omit<Event, 'dates'>
        doorTime?: string | null
        availabilityStarts?: string | null
        subscriberAvailabilityStarts?: string | null
        memberAvailabilityStarts?: string | null
        availabilityEnds?: string | null
        subscriberAvailabilityEnds?: string | null
        memberAvailabilityEnds?: string | null
        preReservationStarts?: string | null
        subscriberPreReservationStarts?: string | null
        memberPreReservationStarts?: string | null
        preReservationEnds?: string | null
        subscriberPreReservationEnds?: string | null
        memberPreReservationEnds?: string | null
        endDate?: string | null
        place?: Omit<Place, 'eventDates'> | null
        bookingLabel?: string | null
        free?: boolean
        minPrice?: number | null
        maxPrice?: number | null
        online?: boolean
        additionalTicketingData?: SecutixAdditionalTicketingData | null
        offers?: AggregateOffer
    }
// ...
    interface Review extends JsonLdObject {
        reviewBody?: string
        author?: string
    }

    interface Seo extends JsonLdObject {
        title?: string | null
        description?: string | null
        keywords?: string | null
    }
// ...
    interface Role extends JsonLdObject {
        name?: string | null
        type?: 'performer' | 'sponsor'
        event?: Event
        people?: Array<Person>
        organization?: Organization
        description?: string | null
        quote?: string | null
        mainRole?: boolean
        onStage?: boolean
        onFocus?: boolean
        manualPeopleOrOrganization?: string
        identities?: string | null
    }

    interface Season extends JsonLdObject, DocumentedEntity, Thing, Routable {
        seo?: Seo
        startDate?: string | null
        endDate?: string | null
        color?: string | null
    }
// ...
    interface ScheduleItem {
        title?: string | null
        headliner?: string | null
        part?: string | null
        duration?: string | null
        intermission?: boolean
    }

    interface KeyValueItem {
        key?: string | null
        value?: string | null
    }
}